pip_requirements.txt
aiapy==0.6.4
astropy==5.0.4
numpy==1.22.3
pandas==1.4.2
sunpy==4.0.3
GitPython==3.1.29
voprov==0.0.2
pydot
